% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_approx.r
\name{boot.approx}
\alias{boot.approx}
\title{Threshold computation from three bootstrap schemes}
\usage{
boot.approx(tau.vec, n.boot, alpha)
}
\arguments{
\item{tau.vec}{vector of tau}

\item{n.boot}{number bootstrap copies}

\item{alpha}{FDR}
}
\value{
a vector of length 3 indicating 3 thresholds
}
\description{
Threshold computation from three bootstrap schemes
}
