% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusmip.r
\name{clusmip_main}
\alias{clusmip_main}
\title{Main function for assessing influential observations by different detection procedures.}
\usage{
clusmip_main(
  X,
  y,
  cluster.type,
  dat.type,
  nfolds,
  n.boot,
  alpha,
  mb.maxComp,
  mp.maxComp,
  selector.switch
)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{cluster.type}{five type of clustering from "kmeans", "kmeans++", "tsne", "spectral" and "rkmeans"}

\item{dat.type}{"gaussian" or "binomial"}

\item{nfolds}{number of folds for cross-validation}

\item{n.boot}{number of bootstrap samples}

\item{alpha}{FDR}

\item{mb.maxComp}{upper bound of number of components for binomial mixture}

\item{mp.maxComp}{upper bound of number of components for Poisson mixture}

\item{selector.switch}{a binary vector of length 4 indicating which selector is used}
}
\value{
A list indicating positions of the potentially influential and clean observations, decision vector from MIP and DFLASSO, and decision from the ClusMIP method for each model selector and approximation method.
}
\description{
Main function for assessing influential observations by different detection procedures.
}
