% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{cluster.est}
\alias{cluster.est}
\title{Clustering}
\usage{
cluster.est(df, nclust, cluster.type)
}
\arguments{
\item{df}{data frame}

\item{nclust}{number of cluster centers}

\item{cluster.type}{types of clustering schemes}
}
\value{
A list of positions of potentially clean and influential observations
}
\description{
Clustering
}
