% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_measure.r
\name{df.measure}
\alias{df.measure}
\title{DF(LASSO)}
\usage{
df.measure(X, Y, dat.type)
}
\arguments{
\item{X}{design matrix}

\item{Y}{response vector}

\item{dat.type}{types of data "gaussian" or "binomial"}
}
\value{
a decision vector indicating the positions of clean and influential observations
}
\description{
DF(LASSO)
}
