% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midquantile_param.r
\name{midcdf.param}
\alias{midcdf.param}
\title{Title}
\usage{
midcdf.param(x, p, param.string, param.val)
}
\arguments{
\item{x}{tau vector}

\item{p}{number of predictors}

\item{param.string}{name of parametric approximating distribution}

\item{param.val}{MLE estimates for parameters}
}
\value{
a midcdf object
}
\description{
Title
}
