% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midquantile_param.r
\name{midquantile.param}
\alias{midquantile.param}
\title{Title}
\usage{
midquantile.param(x, p, param.string, param.val, probs)
}
\arguments{
\item{x}{tau vector}

\item{p}{number of predictors}

\item{param.string}{name of parametric approximating distribution}

\item{param.val}{MLE estimates for parameters}

\item{probs}{percentile}
}
\value{
a midquantile object
}
\description{
Title
}
