% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_approx.r
\name{param.approx}
\alias{param.approx}
\title{Title}
\usage{
param.approx(tau.vec, p, mb.maxComp, mp.maxComp, n.boot, alpha)
}
\arguments{
\item{tau.vec}{tau vector}

\item{p}{number of predictors}

\item{mb.maxComp}{upper bound of number of components for binomial mixture}

\item{mp.maxComp}{upper bound of number of components for Poisson mixture}

\item{n.boot}{number of bootstrap samples for Poisson mixutre order selection (likelihood-based approach)}

\item{alpha}{FDR}
}
\value{
6 thresholds based on the 6 parametric approximating distributions
}
\description{
Title
}
